+++ src/Core/Plugin.php
@@ -32,6 +32,7 @@ use Samsiani\CodeonMultilingual\Url\PostLinkFilter;
 use Samsiani\CodeonMultilingual\Url\Router;
 use Samsiani\CodeonMultilingual\Url\TermLinkFilter;
 use Samsiani\CodeonMultilingual\Woo\ProductSync;
+use Samsiani\CodeonMultilingual\Woo\TranslationLock;
 use Samsiani\CodeonMultilingual\Woo\VariationTranslator;
 
 /**
@@ -81,6 +82,7 @@ final class Plugin {
 		// WooCommerce.
 		ProductSync::register();
 		VariationTranslator::register();
+		TranslationLock::register();
 
 		// Strings.
 		StringTranslator::register();